Feat(TaskCompletion): ajuste front e back para completar a tarefa e d…

+++ todo-frontend/src/contexts/TaskContext.jsx
@@ -28,6 +28,10 @@ export const TaskProvider = ({ children }) => {
     };
 
     const deleteTask = async (id) => {
+        if (!id) {
+            console.error("ID da tarefa está indefinido");
+            return;
+        }
         try {
             await axios.delete(`${process.env.REACT_APP_API_URL}/${id}`);
             setTasks(tasks.filter(task => task.id !== id));
@@ -36,12 +40,24 @@ export const TaskProvider = ({ children }) => {
         }
     };
 
+    const completeTask = async (id, completed) => {
+        try {
+            const response = await axios.patch(`${process.env.REACT_APP_API_URL}/${id}/complete`, { completed });
+            setTasks(tasks.map(task =>
+                task.id === id ? { ...task, completed: response.data.completed } : task
+            ));
+        } catch (error) {
+            console.error("Erro ao completar/descompletar tarefa:", error);
+            throw error;
+        }
+    };
+
     useEffect(() => {
         fetchTasks();
     }, []);
 
     return (
-        <TaskContext.Provider value={{ tasks, loading, addTask, deleteTask }}>
+        <TaskContext.Provider value={{ tasks, loading, addTask, deleteTask, completeTask }}>
             {children}
         </TaskContext.Provider>
     );
